Fix stories
