+++ site/src/components/Resources/AgentRow.stories.tsx
@@ -1,0 +1,35 @@
+import { Story } from "@storybook/react"
+import { MockWorkspace, MockWorkspaceAgent } from "testHelpers/entities"
+import { AgentRow, AgentRowProps } from "./AgentRow"
+
+export default {
+  title: "components/AgentRow",
+  component: AgentRow,
+}
+
+const Template: Story<AgentRowProps> = (args) => <AgentRow {...args} />
+
+export const Example = Template.bind({})
+Example.args = {
+  agent: MockWorkspaceAgent,
+  workspace: MockWorkspace,
+  applicationsHost: "",
+  showApps: true,
+}
+
+export const HideSSHButton = Template.bind({})
+HideSSHButton.args = {
+  agent: MockWorkspaceAgent,
+  workspace: MockWorkspace,
+  applicationsHost: "",
+  showApps: true,
+  hideSSHButton: true,
+}
+
+export const NotShowingApps = Template.bind({})
+NotShowingApps.args = {
+  agent: MockWorkspaceAgent,
+  workspace: MockWorkspace,
+  applicationsHost: "",
+  showApps: false,
+}
